CODE/requirements.txt
fastapi
uvicorn
requests
python-dotenv
python-multipart
pydantic-settings
langchain
langchain-community
langchain-text-splitters
chromadb
sentence-transformers
pypdf2
python-docx
openpyxl
